# backend/assistant-openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Budget AI Assistant",
  "version": "1.2.0",
  "description": "事实查询、写入预览/确认和导出接口。AI 不直接写入预算数据。"
 },
 "servers": [
  {
   "url": "http://127.0.0.1:3760"
  }
 ],
 "security": [
  {
   "AccessToken": []
  }
 ],
 "paths": {
  "/api/assistant/chat": {
   "post": {
    "summary": "自然语言查询或生成操作建议",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "事实、引用和可选操作建议",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatResponse"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/ValidationError"
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/chat/stream": {
   "post": {
    "summary": "SSE 自然语言查询",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "event: token 多次，最后 event: done；done.data 与 ChatResponse 相同并含 done=true",
      "content": {
       "text/event-stream": {}
      }
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    },
    "description": "真流式 SSE。事件顺序：event: open(连接就绪) → event: token(正文增量，多个) → event: error(仅失败时) → event: done(完整 ChatResponse 并附 done:true)。响应头先发出后再转发模型增量，首字延迟等于模型第一个增量；请求体校验在写响应头之前完成，因此参数错误仍返回结构化 400。响应头已发出后的异常以 event: error 下发并补一个 event: done。"
   }
  },
  "/api/assistant/preview": {
   "post": {
    "summary": "创建待确认操作预览",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreviewRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "pending action",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Action"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/ValidationError"
     },
     "409": {
      "$ref": "#/components/responses/ConflictError"
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/actions/{id}/confirm": {
   "post": {
    "summary": "提交确认令牌并执行操作",
    "parameters": [
     {
      "$ref": "#/components/parameters/ActionId"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConfirmRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "confirmed action",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Action"
        }
       }
      }
     },
     "409": {
      "$ref": "#/components/responses/ConflictError"
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/actions/{id}/cancel": {
   "post": {
    "summary": "取消待确认操作",
    "parameters": [
     {
      "$ref": "#/components/parameters/ActionId"
     }
    ],
    "responses": {
     "200": {
      "description": "cancelled/已是终态的 action",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Action"
        }
       }
      }
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/actions/{id}/download": {
   "get": {
    "summary": "下载已确认的导出 action",
    "parameters": [
     {
      "$ref": "#/components/parameters/ActionId"
     }
    ],
    "responses": {
     "200": {
      "description": "CSV 或 XLSX 文件",
      "content": {
       "text/csv": {},
       "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
      }
     },
     "409": {
      "$ref": "#/components/responses/ConflictError"
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
...
  "/api/assistant/conversations/{id}": {
   "get": {
    "summary": "会话及消息",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     }
    ],
    "responses": {
     "200": {
      "description": "conversation detail"
     }
    }
   },
   "patch": {
    "summary": "重命名会话",
    "description": "标题为空时回退为「会话 #id」；写入 operation_log(ai.conversation.rename)。",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "title": {
          "type": "string",
          "maxLength": 200
         }
        },
        "required": [
         "title"
        ]
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "重命名后的会话"
     },
     "404": {
      "description": "会话不存在"
     }
    }
   },
   "delete": {
    "summary": "删除会话",
    "description": "消息随 ON DELETE CASCADE 一起删除；已保存的洞察与已产生的操作只解绑(ON DELETE SET NULL)，仍可追溯。写入 operation_log(ai.conversation.delete)。",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     }
    ],
    "responses": {
     "200": {
      "description": "删除结果，含被删除消息数与被解绑的洞察/操作数",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "id": {
           "type": "integer"
          },
          "messageCount": {
           "type": "integer"
          },
          "insightCount": {
           "type": "integer"
          },
          "actionCount": {
           "type": "integer"
          },
          "pendingActions": {
           "type": "integer"
          }
         }
        }
       }
      }
     },
     "404": {
      "description": "会话不存在"
     }
    }
   }
  },
  "/api/assistant/insights/{id}": {
   "get": {
    "summary": "保存的洞察/依据草稿",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     }
    ],
    "responses": {
     "200": {
      "description": "insight"
     }
    }
   },
   "delete": {
    "summary": "删除洞察",
    "description": "洞察是只读分析快照，删除不影响任何业务数据。写入 operation_log(ai.insight.delete)。",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     }
    ],
    "responses": {
     "200": {
      "description": "删除结果"
     },
     "404": {
      "description": "洞察不存在"
     }
    }
   }
  },
  "/api/assistant/attribution": {
   "post": {
    "summary": "差异归因(只读):按组织、科目和方向排序,支持逐层展开",
    "description": "数字全部来自 completionReport;组织维度与科目维度根层合计必须相等,见 reconciliation。不写任何业务数据。",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AttributionRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "归因结果",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttributionResponse"
        }
       }
      }
     },
     "400": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/report": {
   "post": {
    "summary": "报告生成(只读):执行月报 / 年度复盘 / 预算讨论材料",
    "description": "后端先确定性组稿并附逐节引用;模型可用时只改写叙述文字,数字不变,模型不可用时 narrativeSource=template。不写任何业务数据。",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReportRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "报告草稿",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportResponse"
        }
       }
      }
     },
     "400": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/import-help": {
   "post": {
    "summary": "导入辅助(只读):解释错误、建议组织/科目匹配、列出未匹配和重复项",
    "description": "错误分类与解释为固化规则;匹配建议按编码编辑距离与前缀/名称命中打分,候选一定是库中真实节点。不写任何业务数据。",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImportHelpRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "导入诊断",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ImportHelpResponse"
        }
       }
      }
     },
     "400": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "助手请求超过每分钟上限(AI_RATE_LIMIT_PER_MIN，默认 30)，响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/insights": {
   "get": {
    "summary": "洞察列表",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 200,
       "default": 50
      }
     }
    ],
    "responses": {
     "200": {
      "description": "insight items"
     }
    }
   },
   "post": {
    "summary": "保存分析洞察(后端按参数重新计算,不接受前端数字)",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "kind"
        ],
        "properties": {
         "conversationId": {
          "type": "integer",
          "minimum": 1,
          "nullable": true
         },
         "title": {
          "type": "string",
          "maxLength": 200
         },
         "kind": {
          "type": "string",
          "enum": [
           "execution",
           "attribution",
           "report",
           "anomalies",
           "trend",
           "accuracy",
           "version_variance",
           "historical_comparison",
           "budget_quality"
          ]
         },
         "params": {
          "type": "object",
          "description": "kind=report 时用 params.reportKind 指定报告类型"
         },
         "note": {
          "type": "string",
          "maxLength": 10000
         }
        }
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "已保存的洞察"
     },
     "400": {
      "description": "参数或口径错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/api/assistant/glossary": {
   "get": {
    "summary": "业务解释(字段、状态、金额方向、单位换算、完成率)",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 500
      }
     }
    ],
    "responses": {
     "200": {
      "description": "matched + catalog"
     }
    }
   }
  },
...
  "/api/assistant/quality-advice": {
   "post": {
    "summary": "定稿质量门禁「解释 + 修复建议」(只读)",
    "description": "输入预算版本,返回该版本质量门禁的归并统计、处理顺序与最小修复路径建议。确定性模板先行;模型可用且 NEWFC_QUALITY_AI 未关闭时仅改写行文,数字守卫失败或模型不可用时 advice 回退模板稿(source=template)。不写任何业务数据。叙述生成走独立限流桶(AI_NARRATIVE_RATE_LIMIT_PER_MIN,默认 10)。",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QualityAdviceRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "处理建议",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QualityAdviceResponse"
        }
       }
      }
     },
     "400": {
      "description": "参数错误",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "叙述生成请求超过每分钟上限(AI_NARRATIVE_RATE_LIMIT_PER_MIN,默认 10),响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  },
  "/api/assistant/master-data-semantic-names": {
   "post": {
    "summary": "主数据「语义命名相似」候选对(只读建议)",
    "description": "找出字面距离远但语义可能重名/同义的组织或科目节点对。输入仅为编码+名称清单,不涉金额;输出经白名单校验(编码必须存在于清单),结果永远标记为建议(source=model),不产生任何写入。模型未配置、NEWFC_MASTER_DATA_AI=0 或调用失败时 semanticAvailable=false 且 pairs 为空,体检报告其余部分完整可用。叙述生成走独立限流桶(AI_NARRATIVE_RATE_LIMIT_PER_MIN,默认 10)。",
    "responses": {
     "200": {
      "description": "语义相似候选对",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MasterDataSemanticNamesResponse"
        }
       }
      }
     },
     "429": {
      "description": "叙述生成请求超过每分钟上限(AI_NARRATIVE_RATE_LIMIT_PER_MIN,默认 10),响应含 retryAfterSeconds 与 Retry-After 头"
     }
    }
   }
  }
 },
 "components": {
  "securitySchemes": {
   "AccessToken": {
    "type": "apiKey",
    "in": "header",
    "name": "x-access-token"
   }
  },
  "parameters": {
   "ActionId": {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
     "type": "integer",
     "minimum": 1
    }
   }
  },
  "responses": {
   "ValidationError": {
    "description": "参数校验失败",
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    }
   },
   "ConflictError": {
    "description": "状态、令牌或预览基线冲突",
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    }
   }
  },
  "schemas": {
...
   "ChatRequest": {
    "type": "object",
    "required": [
     "message"
    ],
    "additionalProperties": false,
    "properties": {
     "conversationId": {
      "type": "integer",
      "minimum": 1
     },
     "message": {
      "type": "string",
      "minLength": 1,
      "maxLength": 20000
     },
     "context": {
      "$ref": "#/components/schemas/AssistantContext"
     },
     "pageContext": {
      "$ref": "#/components/schemas/PageContextV2"
     }
    },
    "description": "提问请求。pageContext 为 V2 页面范围快照；提供后优先用于构造页面范围，旧 context 保留兼容。后端先校验 V2，再由本轮问题覆盖页面字段、会话只补空；问题覆盖成功时通过 contextStatus=explicit_override 与 contextTrace.overrides 说明。"
   },
   "ChatResponse": {
    "type": "object",
    "required": [
     "conversationId",
     "text",
     "facts",
     "citations",
     "suggestions",
     "action",
     "navigation",
     "resolvedContext",
     "resolution",
     "routing",
     "model",
     "intents",
     "modelError"
    ],
    "properties": {
     "conversationId": {
      "type": "integer"
     },
     "text": {
      "type": "string"
     },
     "facts": {
      "type": "array",
      "items": {
       "type": "object"
      }
     },
     "citations": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Citation"
      }
     },
     "suggestions": {
      "type": "array",
      "items": {
       "type": "string"
      }
     },
     "action": {
      "nullable": true,
      "allOf": [
       {
        "$ref": "#/components/schemas/ProposedAction"
       }
      ]
     },
     "navigation": {
      "nullable": true,
      "type": "object",
      "description": "页面导航目标(page/path/label/reason/params);无导航意图为 null"
     },
     "resolvedContext": {
      "allOf": [
       {
        "$ref": "#/components/schemas/AssistantContext"
       }
      ],
      "description": "后端最终使用的上下文；前端可据此回填筛选器，下一轮会作为继承上下文"
     },
     "resolution": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/ContextResolution"
      },
      "description": "每一项上下文的来源与依据"
     },
     "routing": {
      "type": "string",
      "enum": [
       "model",
       "rules"
      ],
      "description": "model=模型自主调用只读工具路由；rules=模型不可用或未调用工具时的关键词兜底"
     },
     "model": {
      "type": "string",
      "description": "本轮实际生效的模型名；template 表示模型不可用、答案来自后端确定性模板"
     },
     "intents": {
      "type": "object",
      "description": "关键词意图识别结果。suppressed 是按语义蕴含省略的重复项(父级结果已包含)；inheritedRead 是本轮判定为纯追问时沿用的上一轮只读意图，为空表示意图来自本轮消息",
      "properties": {
       "read": {
        "type": "array",
        "items": {
         "type": "string"
        }
       },
       "write": {
        "type": "array",
        "items": {
         "type": "string"
        }
       },
       "suppressed": {
        "type": "array",
        "items": {
         "type": "string"
        }
       },
       "inheritedRead": {
        "type": "array",
        "items": {
         "type": "string"
        }
       }
      }
     },
     "modelError": {
      "type": "string",
      "nullable": true,
      "description": "模型不可用的原因；null 表示无异常"
     },
     "contextStatus": {
      "type": "string",
      "enum": [
       "aligned",
       "explicit_override"
      ],
      "description": "aligned=已对齐当前页面;explicit_override=问题显式覆盖了页面范围(仅请求带 pageContext 时返回)"
     },
     "effectiveContext": {
      "$ref": "#/components/schemas/EffectiveContext"
     },
     "contextSummary": {
      "type": "string",
      "description": "一行范围摘要,如「预算执行分析 · 2026 年 · 预算 V3 · 江垭电站 · 截至 6 月」;历史会话从 response_json 原样读取(§9.7)"
     },
     "contextTrace": {
      "$ref": "#/components/schemas/ContextTrace"
     },
     "capability": {
      "type": [
       "string",
       "null"
      ],
      "enum": [
       "overview",
       "assistant_content",
       "master_data",
       "budget",
       "actual",
       "execution",
       "comparison",
       "import_conversion",
       "evidence",
       "operations",
       null
      ],
      "description": "本轮实际调用的领域能力;读意图超出 PageCapabilityMap 时被拒绝并如实告知(§9.3)"
     },
     "draftApplied": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DraftAppliedSummary"
       },
       {
        "type": "null"
       }
      ],
      "description": "草稿采用摘要;不含原始修改值"
     }
    }
   },
...
   "PreviewRequest": {
    "type": "object",
    "required": [
     "type"
    ],
    "properties": {
     "type": {
      "type": "string",
      "enum": [
       "budget_draft",
       "scenario",
       "copy_budget",
       "bulk_adjustment",
       "basis_text",
       "export",
       "preview_budget_draft",
       "preview_scenario",
       "preview_copy_budget",
       "preview_bulk_adjustment",
       "preview_basis_text"
      ]
     },
     "conversationId": {
      "type": "integer",
      "minimum": 1
     },
     "idempotencyKey": {
      "type": "string",
      "maxLength": 200
     },
     "params": {
      "type": "object"
     }
    }
   },
   "ConfirmRequest": {
    "type": "object",
    "required": [
     "confirmationToken"
    ],
    "properties": {
     "confirmationToken": {
      "type": "string",
      "minLength": 1
     }
    }
   },
   "Action": {
    "type": "object",
    "required": [
     "id",
     "type",
     "status",
     "preview",
     "expiresAt"
    ],
    "properties": {
     "id": {
      "type": "integer"
     },
     "conversationId": {
      "type": [
       "integer",
       "null"
      ]
     },
     "type": {
      "type": "string"
     },
     "status": {
      "type": "string",
      "enum": [
       "pending",
       "confirmed",
       "cancelled",
       "expired"
      ]
     },
     "preview": {
      "type": "object"
     },
     "expiresAt": {
      "type": "string",
      "format": "date-time"
     },
     "confirmationToken": {
      "type": "string"
     },
     "result": {}
    }
   },
   "Error": {
    "type": "object",
    "required": [
     "code",
     "message"
    ],
    "properties": {
     "code": {
      "type": "string"
     },
     "message": {
      "type": "string"
     },
     "errors": {
      "type": "array"
     },
     "details": {
      "type": "object",
      "description": "安全错误定位信息；V2 错误可含 snapshotId、field 和 reason"
     },
     "retryAfterSeconds": {
      "type": "integer",
      "minimum": 1
     }
    },
    "description": "统一错误结构。V2 相关 code：CONTEXT_INVALID(400，schema/pageKey/白名单/资源或实体格式失败)；CONTEXT_NOT_READY(409，保留码，当前 UI 在发送前阻断未就绪页面)；CONTEXT_STALE(409，verification owner 或 cell focus 来源与当前页面/版本/快照不一致，不由 contextVersion 数字回退触发)；CONTEXT_CONFLICT(409，年度与预算版本/目标版本/实际快照、版本对比年度或树关系矛盾)；CONTEXT_TOO_LARGE(413，非草稿上下文 >64KiB、draft >5MiB、changes >10000、refs >500 或 bounds 列表超限)；DRAFT_STALE(409，草稿 revision/updatedAt/当前批次/版本状态/树基线变化)；CAPABILITY_UNAVAILABLE(400，保留给直接能力调用，普通聊天通常通过能力过滤和 notice 告知)。"
   },
   "AttributionRequest": {
    "type": "object",
    "required": [
     "versionId"
    ],
    "properties": {
     "versionId": {
      "type": "integer",
      "minimum": 1
     },
     "batchId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true,
      "description": "实际快照;缺省按现有取数口径"
     },
     "orgScopeId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true
     },
     "accountScopeId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true
     },
     "sheetKey": {
      "type": "string",
      "nullable": true
     },
     "maxDepth": {
      "type": "integer",
      "minimum": 1,
      "maximum": 10,
      "default": 3,
      "description": "逐层展开层级上限"
     },
     "topN": {
      "type": "integer",
      "minimum": 1,
      "maximum": 100,
      "default": 10,
      "description": "每层保留条目数"
     },
     "direction": {
      "type": "string",
      "enum": [
       "favorable",
       "unfavorable",
       "all"
      ],
      "default": "all"
     }
    }
   },
...
   "AttributionResponse": {
    "type": "object",
    "required": [
     "version",
     "params",
     "totals",
     "byOrg",
     "byAccount",
     "reconciliation"
    ],
    "properties": {
     "version": {
      "type": "object"
     },
     "asOfDate": {
      "type": "string",
      "nullable": true
     },
     "timeProgressValue": {
      "type": "number",
      "nullable": true
     },
     "actualSource": {
      "type": "string",
      "enum": [
       "current",
       "snapshot",
       "final",
       "none"
      ]
     },
     "actualBatchId": {
      "type": "integer",
      "nullable": true
     },
     "treeBasis": {
      "type": "object"
     },
     "scopeBasis": {
      "type": "object"
     },
     "params": {
      "type": "object"
     },
     "totals": {
      "type": "object"
     },
     "byOrg": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/AttributionNode"
      }
     },
     "byAccount": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/AttributionNode"
      }
     },
     "rankedOrgLeaves": {
      "type": "array",
      "items": {
       "type": "object"
      }
     },
     "rankedAccountLeaves": {
      "type": "array",
      "items": {
       "type": "object"
      }
     },
     "quantityVariances": {
      "type": "array",
      "items": {
       "type": "object"
      },
      "description": "数量按 10^4 缩放整数,不参与金额归因"
     },
     "reconciliation": {
      "type": "object",
      "description": "组织维度与科目维度根层/叶子合计核对"
     },
     "notes": {
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
   "ReportRequest": {
    "type": "object",
    "required": [
     "kind"
    ],
    "properties": {
     "kind": {
      "type": "string",
      "enum": [
       "monthly_execution",
       "annual_review",
       "budget_discussion"
      ]
     },
     "versionId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true,
      "description": "月报与讨论材料必填"
     },
     "year": {
      "type": "integer",
      "minimum": 1900,
      "maximum": 9999,
      "nullable": true,
      "description": "年度复盘必填"
     },
     "batchId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true
     },
     "targetVersionId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true,
      "description": "讨论材料的对比版本(同年度)"
     },
     "orgScopeId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true
     },
     "accountScopeId": {
      "type": "integer",
      "minimum": 1,
      "nullable": true
     },
     "sheetKey": {
      "type": "string",
      "nullable": true
     },
     "topN": {
      "type": "integer",
      "minimum": 1,
      "maximum": 50,
      "nullable": true
     }
    }
   },
   "ReportResponse": {
    "type": "object",
    "required": [
     "kind",
     "title",
     "sections",
     "citations",
     "narrative",
     "narrativeSource"
    ],
    "properties": {
     "kind": {
      "type": "string",
      "enum": [
       "monthly_execution",
       "annual_review",
       "budget_discussion"
      ]
     },
     "kindLabel": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "generatedAt": {
      "type": "string"
     },
     "period": {
      "type": "object"
     },
     "scope": {
      "type": "object"
     },
     "sections": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "key": {
         "type": "string"
        },
        "title": {
         "type": "string"
        },
        "bullets": {
         "type": "array",
         "items": {
          "type": "string"
         },
         "description": "确定性要点,金额已按万元两位小数格式化"
        },
        "data": {
         "type": "object",
         "description": "结构化数据,金额仍为整数分"
        },
        "citations": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Citation"
         }
        }
       }
      }
     },
     "facts": {
      "type": "array",
      "items": {
       "type": "object"
      }
     },
     "citations": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Citation"
      }
     },
     "narrative": {
      "type": "string",
      "description": "Markdown 成稿"
     },
     "narrativeSource": {
      "type": "string",
      "enum": [
       "template",
       "model"
      ]
     },
     "model": {
      "type": "string",
      "description": "template 表示模型未启用或已降级"
     },
     "suggestions": {
      "type": "array",
      "items": {
       "type": "string"
      },
      "description": "AI 建议,展示时必须标记为建议"
     },
     "notes": {
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
...
   "QualityAdviceRequest": {
    "type": "object",
    "required": [
     "versionId"
    ],
    "properties": {
     "versionId": {
      "type": "integer",
      "description": "预算版本 ID"
     }
    },
    "additionalProperties": false
   },
...
  }
 }
}